+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     - Fix AdaBoost early stopping algorithm
     - Fix zero quantile calculation
     - Fix Jaccard zero distance on same vectors
+    - Fix Image Rotator zero jitter
 
 - 2.5.3
     - Update PHP stemmer library